move enums into seperate package
