+++ app/src/main/java/com/ottt/ponker/MainActivity.java
@@ -8,6 +8,9 @@ import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.ottt.ponker.enums.Suit;
+import com.ottt.ponker.enums.Tier;
+
 import java.util.ArrayList;
 
 public class MainActivity extends AppCompatActivity {
