+++ app/src/main/java/com/ottt/ponker/enums/Suit.java
@@ -1,4 +1,4 @@
-package com.ottt.ponker;
+package com.ottt.ponker.enums;
 
 public enum Suit {
     HEARTS,
